+++ Model/Plant/Reproductive/GrainPartGN.cpp
@@ -259,7 +259,7 @@ void fruitGrainPartGN::doNDemandGrain (float nfact_grain_conc      //   (INPUT)
       // during grain C filling period so make sure that C filling is still
       // going on otherwise stop putting N in now
 
-      grain_growth = divide((mealPart->Growth.DM() + mealPart->Retranslocation.DM())
+      grain_growth = (float)divide((mealPart->Growth.DM() + mealPart->Retranslocation.DM())
                             , gGrain_no
                             , 0.0);
       if (grain_growth < cCrit_grainfill_rate)
